+++ NEWS.md
@@ -3,6 +3,11 @@ title: "NEWS"
 output: html_document
 ---
 
+## v1.2.3
+
+* bugfix in RIV point calculation
+
+
 ## v1.2.2
 
 * added 2025 Journal Citation Index dataset.
@@ -21,9 +26,9 @@ output: html_document
 
 ## v1.1.0
 
-* added 2023 Journal Citation Index dataset.
+* added 2023 Journal Citation Index dataset
 
 
 ## v1.0.0
 
-* initial version.
+* initial version